Use single `NameAndImageExtraData` for both Channel and User models
